+++ changelog.md
@@ -5,6 +5,7 @@
 - Removes deprecated concepts (e.g. enterprise and software system/person location).
 - structurizr-core: Adds `Workspace.trim()` to trim a workspace of unused elements (i.e. those not associated with any views).
 - structurizr-core: Adds support for SVG image views (https://github.com/structurizr/java/issues/249).
+- structurizr-core: View keys will be automatically generated if not specified.
 - structurizr-client: Removes `StructurizrClient` (use `WorkspaceApiClient` instead).
 - structurizr-import: Adds support for importing decisions managed by Log4brains.
 - structurizr-import: Adds support for importing decisions in MADR format.